+++ gateway/augment.go
@@ -1,3 +1,5 @@
+// TODO_UPNEXT(@adshmh): Add a mermaid diagram of the different structural
+// (i.e. packages, types) components to help clarify the role of each.
 package gateway
 
 import (
@@ -8,7 +10,17 @@ import (
 	"github.com/buildwithgrove/path/relayer"
 )
 
-// Protocol specifies the interactions of the EndpointDataAugmenter with
+// endpointHydratorRunIntervalMillisec specifies the running
+// interval of an endpoint hydrator.
+const endpointHydratorRunIntervalMillisec = 30_000
+
+// TODO_UPNEXT(@adshmh): Complete the following to remove the confusing Protocol interface below:
+//
+//	1- Split the relayer package's Protocol interface.
+//	2- Import the appropriate interface here, e.g. a new `EndpointProvider` interface.
+//	3- Update/remove the comment below.
+//
+// Protocol specifies the interactions of the EndpointHydrator with
 // the underlying protocol.
 // It is defined separately, rather than reusing relayer.Protocol interface,
 // to ensure only minimum necessary capabilities are available to the augmenter.
@@ -16,7 +28,9 @@ type Protocol interface {
 	Endpoints(relayer.ServiceID) ([]relayer.Endpoint, error)
 }
 
-// EndpointDataAugmenter augments the available dataset on quality of endpoints.
+// EndpointHydrator augments the available dataset on quality of endpoints.
+// Please see the following link for details on the Hydrator section of the name:
+// https://stackoverflow.com/questions/6991135/what-does-it-mean-to-hydrate-an-object
 // It does this to ensure each supported service's QoS instance has enough data
 // on each available endpoint to make an informed selection of an endpoint
 // to handle a user service request.
@@ -26,14 +40,16 @@ type Protocol interface {
 // 2. Performing the required checks on the endpoint, in the form
 // of a (synthetic) service request.
 // 3. Reporting the results back to the service's QoS instance.
-type EndpointDataAugmenter struct {
+type EndpointHydrator struct {
 	Protocol
 	*relayer.Relayer
 	QoSPublisher
-	Services map[relayer.ServiceID]QoSEndpointCheckGenerator
+	ServiceQoSGenerators map[relayer.ServiceID]QoSEndpointCheckGenerator
 }
 
-func (eda *EndpointDataAugmenter) Start() error {
+// Start should be called to signal this instance of the hydrator
+// to start generating and sending out the endpoint check requests.
+func (eda *EndpointHydrator) Start() error {
 	if eda.Protocol == nil {
 		return errors.New("a Protocol instance must be proivded.")
 	}
@@ -42,13 +58,17 @@ func (eda *EndpointDataAugmenter) Start() error {
 		return errors.New("a Relayer must be provided.")
 	}
 
-	if len(eda.Services) == 0 {
+	if eda.QoSPublisher == nil {
+		return errors.New("a QoS Publisher must be provided.")
+	}
+
+	if len(eda.ServiceQoSGenerators) == 0 {
 		return errors.New("at-least one covered service must be specified")
 	}
 
 	go func() {
 		// TODO_IMPROVE: support configuring a custom running interval.
-		ticker := time.NewTicker(30 * time.Second)
+		ticker := time.NewTicker(endpointHydratorRunIntervalMillisec * time.Millisecond)
 		for {
 			eda.run()
 			<-ticker.C
@@ -58,8 +78,8 @@ func (eda *EndpointDataAugmenter) Start() error {
 	return nil
 }
 
-func (eda *EndpointDataAugmenter) run() {
-	for svcID, svcQoS := range eda.Services {
+func (eda *EndpointHydrator) run() {
+	for svcID, svcQoS := range eda.ServiceQoSGenerators {
 		go func(serviceID relayer.ServiceID, serviceQoS QoSEndpointCheckGenerator) {
 			eda.performChecks(serviceID, serviceQoS)
 		}(svcID, svcQoS)
@@ -68,27 +88,29 @@ func (eda *EndpointDataAugmenter) run() {
 	// TODO_IMPROVE: wait for all goroutines to finish before returning.
 }
 
-func (eda *EndpointDataAugmenter) performChecks(serviceID relayer.ServiceID, serviceQoS QoSEndpointCheckGenerator) {
-	// endpoints here is expected to be: []Endpoint (AppAddr and EndpointAddr should be properties of Endpoint interface)
+func (eda *EndpointHydrator) performChecks(serviceID relayer.ServiceID, serviceQoS QoSEndpointCheckGenerator) {
 	endpoints, err := eda.Protocol.Endpoints(serviceID)
 	if err != nil {
-		// TODO_IMPROVE: log the error
+		// TODO_IN_THIS_COMMIT: log the error
 		return
 	}
 
 	// TODO_FUTURE: use a single goroutine per endpoint
 	for _, endpoint := range endpoints {
 		endpointAddr := endpoint.Addr()
-		endpointChecks := serviceQoS.GetRequiredQualityChecks(endpointAddr)
-		if len(endpointChecks) == 0 {
-			// TODO_FUTURE: Log an info-level message
+		requiredChecks := serviceQoS.GetRequiredQualityChecks(endpointAddr)
+		if len(requiredChecks) == 0 {
+			// TODO_IN_THIS_COMMIT: Log an info-level message
 			continue
 		}
 
-		for _, serviceRequestCtx := range endpointChecks {
-			// TODO_IMPROVE: Sending a request here should use some method shared with the user request handler.
+		for _, serviceRequestCtx := range requiredChecks {
+			// TODO_IMPROVE: Sending a request here should use some method shared with
+			// the user request (i.e. HTTP request) handler.
 			// This would ensure that both organic, i.e. user-generated, and quality data augmenting service requests
 			// take the same execution path.
+			// TODO_UPNEXT(@adshmh): remove the context input argument once the Relayer interface's
+			// SendRelay function is updated.
 			endpointResponse, err := eda.Relayer.SendRelay(
 				context.TODO(),
 				serviceID,
@@ -96,14 +118,19 @@ func (eda *EndpointDataAugmenter) performChecks(serviceID relayer.ServiceID, ser
 				serviceRequestCtx.GetEndpointSelector(),
 			)
 
-			// Protocol-level errors are the responsibility of the specific
-			// protocol instance used in serving the request.
-			// e.g. an endpoint that is temporarily/permanently unavailable
-			// should not be returned by the AvailableEndpoints() method.
+			// Ignore any errors returned from the SendRelay call above.
+			// These would be protocol-level errors, which are the responsibility
+			// of the specific protocol instance used in serving the request.
+			// e.g. the Protocol instance should drop an endpoint that is
+			// temporarily/permanently unavailable from the set returned by
+			// the Endpoints() method.
+			//
+			// There is no action required from the QoS perspective, if no
+			// responses were received from an endpoint.
 			if err != nil {
 				// TODO_FUTURE: consider retrying failed service requests
 				// as the failure may not be related to the quality of the endpoint.
-				// TODO_IMPROVE: log the error
+				// TODO_IN_THIS_COMMIT: log the error
 				continue
 			}
 
@@ -112,7 +139,7 @@ func (eda *EndpointDataAugmenter) performChecks(serviceID relayer.ServiceID, ser
 			// TODO_FUTURE: consider supplying additional data to QoS.
 			// e.g. data on the latency of an endpoint.
 			if err := eda.QoSPublisher.Publish(serviceRequestCtx.GetObservationSet()); err != nil {
-				// TODO_IMPROVE: log the error
+				// TODO_IN_THIS_COMMIT: log the error
 			}
 		}
 	}
