+++ message/observe.go
@@ -21,7 +21,7 @@ type ObservationSet interface {
 	Broadcast() error
 }
 
-// TODO_FUTURE: consider using protobuf.
+// TODO_UPNEXT(@adshmh): consider using protobuf.
 // Unmarshaller builds an instance of the ObservationSet,
 // matching a specific implementation, e.g. one provided by a service QoS.
 // Each service's QoS instance provides its own unique unmarshaller.
